task13-

+++ app/src/main/java/pl/futurecollars/invoicing/db/file/FileBasedDatabase.java
@@ -21,7 +21,7 @@ public class FileBasedDatabase implements Database {
     private final JsonService jsonService;
 
     @Override
-    public int save(Invoice invoice) {
+    public long save(Invoice invoice) {
         try {
             invoice.setId(idService.getNextId());
             fileService.appendLineToFile(invoicesPath, jsonService.objectToJson(invoice));
@@ -32,7 +32,7 @@ public class FileBasedDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> getById(int id) {
+    public Optional<Invoice> getById(long id) {
         try {
             return fileService.readAllLines(invoicesPath)
                     .stream()
@@ -57,7 +57,7 @@ public class FileBasedDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+    public Optional<Invoice> update(long id, Invoice updatedInvoice) {
         Optional<Invoice> toUpdate = getById(id);
 
         updatedInvoice.setId(id);
@@ -79,7 +79,7 @@ public class FileBasedDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> delete(int id) {
+    public Optional<Invoice> delete(long id) {
         Optional<Invoice> toDelete = getById(id);
         try {
             List<String> updatedInvoicesList = fileService.readAllLines(invoicesPath)
@@ -95,7 +95,7 @@ public class FileBasedDatabase implements Database {
         }
     }
 
-    private boolean containsId(String line, int id) {
+    private boolean containsId(String line, long id) {
         return line.contains("{\"id\":" + id + ",\"number\"");
     }
 }
